refactor:优化客户端新注册逻辑
